% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{binomial plot}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{the number of iterations}

\item{n}{the size of the sample}

\item{p}{Probability of success}
}
\value{
a plot
}
\description{
binomial plot
}
\examples{
\dontrun{iter = 50, n = 30, q = 0.7; mybin(50, 30, 0.7)}

}
